Complete refactoring of the send queue, Groups new feature
now it is possible to browse the outbox from mutt and even
delete mails that were meant to be sent. This also introduces
a new feature: Groups is a directory containing simple files
with recipient addresses one per line: they are basically
one-way mailout lists and the main feature is that every email
sent to each recipient will have it as unique To: definitely
avoiding huge Cc: lists but also Bcc: blacklisting on SMTP etc.