changed the way passwords are stored for better usability
now there aren't different passwords for different protocols on same hosts,
we assume a user has same password for imap and pop and smtp on same host.
This will force all users to retype a password and save a new record in the
keyring they are using, because also the password hashing scheme changed.