Command redesign to leverage pipe operations
This is a rather big refactoring of the way commanands work and
interoperate, so that one can pipe inputs and outputs of commands
efficiently to achieve more complex operations, i.e: take the results of
a search command, extract all email addresses found in them and export
all as a vcard.

Documentation in the user manual is updated accordingly. This
refactoring also eliminates quite some duplicate and obsolete code.